+++ lib/compony/component.rb
@@ -111,16 +111,6 @@ module Compony
       comp_cst.to_s.underscore
     end
 
-    # @todo deprecate (check for usages beforehand)
-    def comp_class_for(...)
-      Compony.comp_class_for(...)
-    end
-
-    # @todo deprecate (check for usages beforehand)
-    def comp_class_for!(...)
-      Compony.comp_class_for!(...)
-    end
-
     # DSL method
     def before_render(&block)
       @before_render_block = block