+++ lib/compony/component.rb
@@ -129,8 +129,7 @@ module Compony
     # Do not overwrite.
     def render(controller, standalone: false, **locals)
       # Call before_render hook if any and backfire instance variables back to the component
-      # TODO: Can .request_context be removed from the next line? Test well!
-      RequestContext.new(self, controller, locals:).request_context.evaluate_with_backfire(&@before_render_block) if @before_render_block
+      RequestContext.new(self, controller, locals:).evaluate_with_backfire(&@before_render_block) if @before_render_block
       # Render, unless before_render has already issued a body (e.g. through redirecting).
       if controller.response_body.nil?
         fail "#{self.class.inspect} must define `content` or set a response body in `before_render`" if @content_blocks.none?